+++ app/src/components/molecules/Form/AutoComplete/index.tsx
@@ -1,16 +1,8 @@
 import React, { ReactElement } from 'react'
-import Select, { ActionMeta } from 'react-select'
+import Select from 'react-select'
 import ErrorMessage from '../../../atoms/ErrorMessage'
 import StyledLabel from './styled'
-
-type AutoCompleteProps = {
-  name: string
-  labelText: string
-  options: Option[] | undefined
-  handleOnChange: (option: Option | null, actionMeta: ActionMeta<Option>) => void
-  setRef?: any
-  errors: errorObject
-}
+import AutoCompleteProps from './type'
 
 const AutoComplete: React.FC<AutoCompleteProps> = ({
   name,
@@ -40,10 +32,6 @@ const AutoComplete: React.FC<AutoCompleteProps> = ({
       {errors && <ErrorMessage errorObject={errors} />}
     </StyledLabel>
   )
-}
-
-AutoComplete.defaultProps = {
-  setRef: {},
 }
 
 export default AutoComplete
