+++ app/src/constants/REPLACEMENT_IMAGES.ts
@@ -1,7 +1,6 @@
 // these are 'mocked' images that serve as placeholders when there is no content
 export default {
-  heroBanner:
-    'https://res.cloudinary.com/dqnks1cyu/image/upload/v1664962461/recipes/food_ytjrov.jpg',
+  heroBanner: 'https://res.cloudinary.com/dqnks1cyu/image/upload/v1664962461/recipes/food_ytjrov.jpg',
   recipeCard:
     'https://res.cloudinary.com/dqnks1cyu/image/upload/v1664962512/recipes/healthy-eating-ingredients-732x549-thumbnail_y5ier5.jpg',
 }
