+++ app/src/redux/reducers/recipes/helpers.ts
@@ -1,9 +1,6 @@
 // unit test required
 
-const replaceRecipeWithIdInArrayWithRecipes = (
-  recipes: Recipe[],
-  updatedRecipe: Recipe,
-): Recipe[] => {
+const replaceRecipeWithIdInArrayWithRecipes = (recipes: Recipe[], updatedRecipe: Recipe): Recipe[] => {
   return recipes.map((recipe: Recipe) => {
     return recipe.id === updatedRecipe.id ? updatedRecipe : recipe
   })
