+++ app/src/redux/reducers/recipes/thunks/deleteRecipe.ts
@@ -2,18 +2,12 @@ import { createAsyncThunk } from '@reduxjs/toolkit'
 import axios from 'axios'
 import RootState from '../../../../types/RootState'
 
-async function deleteRecipeAPI(
-  data: { id: number; authorId: string },
-  token: string,
-): Promise<Recipe> {
-  const response = await axios.delete(
-    `${process.env.REACT_APP_SERVER_URL}/api/recipes/${data.id}`,
-    {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
+async function deleteRecipeAPI(data: { id: number; authorId: string }, token: string): Promise<Recipe> {
+  const response = await axios.delete(`${process.env.REACT_APP_SERVER_URL}/api/recipes/${data.id}`, {
+    headers: {
+      Authorization: `Bearer ${token}`,
     },
-  )
+  })
   return response.data
 }
